+++ lib/src/concurrency-controller/create-multi-concurrency-controller.spec.ts
@@ -60,7 +60,7 @@ describe('createMultiConcurrencyController', () => {
     // Arrange
     const controller = createMultiConcurrencyController(
       () => 'discriminating-mutex',
-      (message) => message.messageType,
+      { discriminator: (message) => message.messageType },
     );
     const orderA: number[] = [];
     const orderB: number[] = [];
@@ -141,6 +141,51 @@ describe('createMultiConcurrencyController', () => {
     expect(new Date().getTime() - start).toBeGreaterThanOrEqual(40);
   });
 
+  it('calls and finishes tasks in the correct order when the semaphore is selected', async () => {
+    // Arrange
+    const controller = createMultiConcurrencyController(() => 'semaphore', {
+      maxSemaphoreParallelism: 2,
+    });
+    const order: number[] = [];
+    const firstTask = async () => {
+      await sleep(30);
+      order.push(1);
+    };
+    const secondTask = async () => {
+      await sleep(10);
+      order.push(2);
+    };
+    const thirdTask = async () => {
+      await sleep(10);
+      order.push(3);
+    };
+    const start = new Date().getTime();
+
+    // Act - these will execute in parallel, but the mutex should ensure they complete in order.
+    await Promise.all([
+      protectedAsyncFunction(
+        controller,
+        firstTask,
+        createOutboxMessage(1, 'A'),
+      ),
+      protectedAsyncFunction(
+        controller,
+        secondTask,
+        createOutboxMessage(1, 'A'),
+      ),
+      protectedAsyncFunction(
+        controller,
+        thirdTask,
+        createOutboxMessage(1, 'A'),
+      ),
+    ]);
+
+    // Assert - verify the order of execution
+    expect(order).toEqual([2, 3, 1]);
+    expect(new Date().getTime() - start).toBeGreaterThanOrEqual(30);
+    expect(new Date().getTime() - start).toBeLessThan(40);
+  });
+
   it('Executes tasks in correct order when different concurrency types are combined', async () => {
     // Arrange
     const controller = createMultiConcurrencyController(
@@ -150,15 +195,23 @@ describe('createMultiConcurrencyController', () => {
             return 'discriminating-mutex';
           case 'B':
             return 'full-concurrency';
-          default:
+          case 'C':
             return 'mutex';
+          case 'D':
+            return 'semaphore';
+          default:
+            throw new Error('unreachable');
         }
       },
-      (message) => message.aggregateType,
+      {
+        discriminator: (message) => message.aggregateType,
+        maxSemaphoreParallelism: 2,
+      },
     );
     const orderA: number[] = [];
     const orderB: number[] = [];
     const orderC: number[] = [];
+    const orderD: number[] = [];
     const createTask =
       (orderArray: number[], sleepTime: number) =>
       async (message: OrderMessage) => {
@@ -214,6 +267,22 @@ describe('createMultiConcurrencyController', () => {
         createTask(orderC, 20),
         createOutboxMessage(8, 'C'),
       ),
+      // semaphore tasks (9 and 10 then 11 --> takes ~40ms)
+      protectedAsyncFunction(
+        controller,
+        createTask(orderD, 20),
+        createOutboxMessage(9, 'D'),
+      ),
+      protectedAsyncFunction(
+        controller,
+        createTask(orderD, 30),
+        createOutboxMessage(10, 'D'),
+      ),
+      protectedAsyncFunction(
+        controller,
+        createTask(orderD, 20),
+        createOutboxMessage(11, 'D'),
+      ),
     ]);
 
     // Assert - verify that different strategies run in parallel and within in the desired concurrency logic
@@ -227,6 +296,8 @@ describe('createMultiConcurrencyController', () => {
     expect(orderB).toContain(6);
 
     expect(orderC).toEqual([7, 8]);
+
+    expect(orderD).toEqual([9, 10, 11]);
 
     expect(diff).toBeGreaterThanOrEqual(40);
     expect(diff).toBeLessThan(60);
@@ -257,11 +328,18 @@ describe('createMultiConcurrencyController', () => {
             return 'discriminating-mutex';
           case 'B':
             return 'full-concurrency';
-          default:
+          case 'C':
             return 'mutex';
+          case 'D':
+            return 'semaphore';
+          default:
+            throw new Error('unreachable');
         }
       },
-      (message) => message.aggregateType,
+      {
+        discriminator: (message) => message.aggregateType,
+        maxSemaphoreParallelism: 2,
+      },
     );
     const success: number[] = [];
     const error: number[] = [];
@@ -286,7 +364,7 @@ describe('createMultiConcurrencyController', () => {
       createOutboxMessage(2, 'A', 'D1'),
     ).catch(() => error.push(2));
 
-    // concurrent tasks (5+6 in parallel --> takes ~40ms)
+    // concurrent tasks (3+4 in parallel)
     protectedAsyncFunction(
       controller,
       createTask(success, 5),
@@ -298,7 +376,7 @@ describe('createMultiConcurrencyController', () => {
       createOutboxMessage(4, 'B'),
     ).catch(() => error.push(4));
 
-    // mutex tasks (7 then 8 --> takes ~40ms)
+    // mutex tasks (5 then 6)
     protectedAsyncFunction(
       controller,
       createTask(success, 5),
@@ -310,11 +388,29 @@ describe('createMultiConcurrencyController', () => {
       createOutboxMessage(6, 'C'),
     ).catch(() => error.push(6));
 
+    // semaphore tasks (9 and 10 in parallel then 11)
+    protectedAsyncFunction(
+      controller,
+      createTask(success, 5),
+      createOutboxMessage(7, 'D'),
+    ).catch(() => error.push(7));
+    protectedAsyncFunction(
+      controller,
+      createTask(success, 5),
+      createOutboxMessage(8, 'D'),
+    ).catch(() => error.push(8));
+    protectedAsyncFunction(
+      controller,
+      createTask(success, 50),
+      createOutboxMessage(9, 'D'),
+    ).catch(() => error.push(9));
+
+    await sleep(5);
     controller.cancel();
 
     // Assert
     await sleep(100);
-    expect(success.sort()).toEqual([1, 3, 4, 5]);
-    expect(error.sort()).toEqual([2, 6]);
+    expect(success.sort()).toEqual([1, 3, 4, 5, 7, 8]);
+    expect(error.sort()).toEqual([2, 6, 9]);
   });
 });